BotChangeViewAngles: float FPU temps remove 8-byte align prologue (-99 byte_diffs)
locals gone, MSVC6 emits sub esp,0x8 without the and esp,0xfffffff8
prologue — matching the reference at 0x10029150.

Trade-off: comparison constants like `v2 > 180.0` now compare a float
against a double literal, which MSVC widens via fld then fcomp QWORD.
The ref does the same with the float load + double-cmp idiom, so this
matches exactly.

audit: MATCH=480 ICM=37 byte_diffs=165687 (was 165786, -99)
aarch64 build OK.

Source-commit: 4f61e5fc85c12fa7cb6e933e3e6dc76699893ba7

+++ botlib/botlib.c
@@ -24231,12 +24231,12 @@ double BotChangeViewAngle(float a1, float a2, float a3)
  * bs->thinktime from the struct.  Leave `thinktime` unused on purpose. */
 int __cdecl BotChangeViewAngles(bot_state_t *bs, float thinktime)
 {
-  double v2; // st7
+  float v2; // st7 (was double)
   float *v3; // esi
   int v4; // ebx
-  double v5; // st7
-  double v6; // st6
-  double v7; // st6
+  float v5; // st7 (was double)
+  float v6; // st6 (was double)
+  float v7; // st6 (was double)
   float v9; // [esp+10h] [ebp-4h]
   float v10; // [esp+18h] [ebp+4h]
   float v11; // [esp+18h] [ebp+4h]
